refactor: 드래그 스크롤 로직 공통 훅 분리 및 FilterSortBar 리팩토링 #7
